# doc/api.yml
openapi: 3.0.3
info:
  title: HEI - Graduate Management API
  description: >
    API for managing grades, groups, tracks and graduates at HEI.
    Roles: STUDENT (read own grades only), TEACHER (read/write grades of
    students they actually teach), ADMIN (full access).
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /students/{studentId}/transcript/send:
    parameters:
      - $ref: '#/components/parameters/StudentId'
    post:
      tags: [Transcripts]
      summary: Trigger asynchronous PDF transcript generation and email delivery
      description: STUDENT (self), TEACHER (ownership), or ADMIN. Asynchronous processing (@Async), immediate 202 response.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [academicYear]
              properties:
                academicYear: {type: string, example: "2024-2025"}
      responses:
        '202':
          description: Accepted, delivery in progress in background
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TranscriptDelivery'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/users:
    get:
      tags: [Admin - Users]
      summary: List users
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/User'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Admin - Users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserCreateRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      tags: [Admin - Users]
      summary: User detail
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Admin - Users]
      summary: Update a user
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserUpdateRequest'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Admin - Users]
      summary: Deactivate a user (soft delete)
      description: >
        Implemented as a soft delete (active=false field), never a real SQL deletion,
        to preserve the integrity of related historical GRADE/COURSE_ASSIGNMENT_TEACHING records.
      responses:
        '204': {description: Deactivated}
        '409':
          description: "Blocked: related historical data exists"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    StudentId:
      name: studentId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  schemas:
    Role:
      type: string
      enum: [STUDENT, TEACHER, ADMIN]
...
    DeliveryStatus:
      type: string
      enum: [IN_PROGRESS, SENT, FAILED]
    User:
      type: object
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        role: {$ref: '#/components/schemas/Role'}
    UserCreateRequest:
      type: object
      required: [email, password, role]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
        role: {$ref: '#/components/schemas/Role'}
    UserUpdateRequest:
      type: object
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
        role: {$ref: '#/components/schemas/Role'}
...
    TranscriptDelivery:
      type: object
      properties:
        id: {type: string, format: uuid}
        studentId: {type: string, format: uuid}
        academicYear: {type: string}
        sentAt: {type: string, format: date-time}
        s3Url: {type: string, format: uri}
        status: {$ref: '#/components/schemas/DeliveryStatus'}
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Resource not found
        path:
          type: string
          example: /students/123/grades
  responses:
    BadRequest:
      description: Invalid request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'